requirements.txt
# ============================================
# Python Dependencies for Food Recognition AI
# ============================================
# Проект: Распознавание блюд с помощью ResNet50
# Установка: pip3 install -r requirements.txt
# ============================================

# --- Core ML/AI Libraries ---
tensorflow==2.18.0              # Deep learning framework для обучения и инференса модели
tensorflow-datasets==4.9.6      # Загрузка датасета Food-101 (101 класс, 75,750 изображений)
keras==3.7.0                    # High-level API для построения нейронных сетей

# --- Image Processing ---
Pillow==11.0.0                  # Загрузка и обработка изображений (PIL)
numpy==1.26.4                   # Операции с массивами (preprocessing изображений)

# --- Web Server (Flask API) ---
Flask==3.1.0                    # Микрофреймворк для REST API сервера
Flask-CORS==5.0.0               # Поддержка CORS для кросс-доменных запросов

# --- Data Science & Visualization ---
scikit-learn==1.5.2             # Утилиты для ML (train/test split, метрики)
matplotlib==3.9.2               # Визуализация графиков обучения
seaborn==0.13.2                 # Статистические графики и heatmaps

# --- Protocol Buffers ---
protobuf>=5.28.0                # Сериализация данных (требуется для TensorFlow 2.18+)

# --- Optional Dependencies ---
# openai==1.54.0                # OpenAI API client для ChatGPT (раскомментируйте если нужно)
# opencv-python==4.8.0          # Дополнительная обработка изображений (не используется в текущей версии)
# pandas==2.0.0                 # Работа с табличными данными (не используется в текущей версии)


# ============================================
# Версии библиотек
# ============================================
# Текущая конфигурация протестирована на:
# - Python 3.11
# - macOS (Apple Silicon M1/M2)
# - TensorFlow 2.18.0 с поддержкой Metal (GPU acceleration)
#
# Точность модели: 81.2% (best_model.keras)
# Размер модели: 222 MB
# ============================================
